+++ internal/acceptance_tests/blocks/logging_kafka_nested.tf
@@ -1,0 +1,5 @@
+logging_kafka {
+  name    = "{{.LOGGING_KAFKA_NAME}}"
+  brokers = "127.0.0.1:9092"
+  topic   = "test-topic"
+}